Replace wrong template selection, with true default
