--- docs/storage/snippets/tutorial/AspireStorage/AspireStorage/AspireStorage.Client/Program.cs
@@ -1,5 +1,0 @@
-using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
-
-var builder = WebAssemblyHostBuilder.CreateDefault(args);
-
-await builder.Build().RunAsync();
