--- docs/storage/snippets/tutorial/AspireStorage/AspireStorage/AspireStorage.Client/wwwroot/appsettings.json
@@ -1,8 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  }
-}
