--- docs/storage/snippets/tutorial/AspireStorage/AspireStorage/AspireStorage/AspireStorage.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\AspireStorage.ServiceDefaults\AspireStorage.ServiceDefaults.csproj" />
-    <ProjectReference Include="..\AspireStorage.Client\AspireStorage.Client.csproj" />
-    <PackageReference Include="Aspire.Azure.Storage.Blobs" Version="8.0.0-preview.1.23557.2" />
-    <PackageReference Include="Aspire.Azure.Storage.Queues" Version="8.0.0-preview.1.23557.2" />
-    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Server" Version="8.0.0" />
-  </ItemGroup>
-
-</Project>
