--- docs/storage/snippets/tutorial/AspireStorage/AspireStorage/AspireStorage/Components/Pages/Home.razor
@@ -1,58 +1,0 @@
-﻿@page "/"
-
-@using Azure.Storage.Blobs
-@using Azure.Storage.Queues
-
-@inject BlobServiceClient BlobClient
-@inject QueueServiceClient QueueServiceClient
-
-<PageTitle>Home</PageTitle>
-
-<div class="text-center">
-    <h1 class="display-4">Request Support</h1>
-</div>
-
-<EditForm Model="@Ticket" FormName="Tickets" method="post" 
-    OnValidSubmit="@HandleValidSubmit" enctype="multipart/form-data">
-    <div class="mb-4">
-        <label>Issue Title</label>
-        <InputText class="form-control" @bind-Value="@Ticket.Title" />
-    </div>
-    <div class="mb-4">
-        <label>Issue Description</label>
-        <InputText class="form-control" @bind-Value="@Ticket.Description" />
-    </div>
-    <div class="mb-4">
-        <label>Attachment</label>
-        <InputFile class="form-control" name="Ticket.Document" />
-    </div>
-    <button class="btn btn-primary" type="submit">Submit</button>
-</EditForm>
-
-@code {
-    [SupplyParameterFromForm(FormName = "Tickets")]
-    private SupportTicket Ticket { get; set; } = new();
-
-    private async Task HandleValidSubmit()
-    {
-        var docsContainer = BlobClient.GetBlobContainerClient("fileuploads");
-
-        // Upload file to blob storage
-        await docsContainer.UploadBlobAsync(
-            Ticket.Document.FileName,
-            Ticket.Document.OpenReadStream());
-
-        // Send message to queue
-        var queueClient = QueueServiceClient.GetQueueClient("tickets");
-
-        await queueClient.SendMessageAsync(
-             $"{Ticket.Title} - {Ticket.Description}");
-    }
-
-    private class SupportTicket()
-    {
-        public string Title { get; set; }
-        public string Description { get; set; }
-        public IFormFile Document { get; set; }
-    }
-}
