--- docs/storage/snippets/tutorial/AspireStorage/AspireStorage/AspireStorage/appsettings.json
@@ -1,13 +1,0 @@
-{
-    "Logging": {
-        "LogLevel": {
-            "Default": "Information",
-            "Microsoft.AspNetCore": "Warning"
-        }
-    },
-    "AllowedHosts": "*",
-    "ConnectionStrings": {
-        "BlobConnection": "https://{account_name}.blob.core.windows.net/",
-        "QueueConnection": "https://{account_name}.queue.core.windows.net/"
-    }
-}
