+++ docs/storage/snippets/tutorial/AspireStorage/AspireStorage/Components/App.razor
