+++ docs/storage/snippets/tutorial/AspireStorage/AspireStorage/Components/Layout/MainLayout.razor
