+++ docs/storage/snippets/tutorial/AspireStorage/AspireStorage/Components/Layout/NavMenu.razor
