+++ docs/storage/snippets/tutorial/AspireStorage/AspireStorage/Components/Pages/Counter.razor
@@ -1,5 +1,5 @@
 ﻿@page "/counter"
-@rendermode InteractiveAuto
+@rendermode InteractiveServer
 
 <PageTitle>Counter</PageTitle>
