+++ docs/storage/snippets/tutorial/AspireStorage/AspireStorage/Components/Pages/Home.razor
@@ -1,0 +1,7 @@
+﻿@page "/"
+
+<PageTitle>Home</PageTitle>
+
+<h1>Hello, world!</h1>
+
+Welcome to your new app.
