+++ docs/storage/snippets/tutorial/AspireStorage/AspireStorage/Components/Pages/Weather.razor
