+++ docs/storage/snippets/tutorial/AspireStorage/AspireStorage/Components/_Imports.razor
@@ -7,5 +7,4 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
 @using AspireStorage
-@using AspireStorage.Client
 @using AspireStorage.Components
