+++ docs/storage/snippets/tutorial/AspireStorage/AspireStorage/appsettings.Development.json
