+++ docs/storage/snippets/tutorial/AspireStorage/AspireStorage/appsettings.json
@@ -4,5 +4,6 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
-  }
+  },
+  "AllowedHosts": "*"
 }
