+++ docs/storage/snippets/tutorial/AspireStorage/AspireStorage/wwwroot/app.css
