+++ docs/storage/snippets/tutorial/AspireStorage/AspireStorage/wwwroot/favicon.png
